For production

+++ app/Services/EmployeeService.php
@@ -4,37 +4,69 @@ namespace App\Services;
 
 use App\Contracts\Import;
 use App\Contracts\Repository;
+use App\Events\EmployeesImported;
+use Carbon\Carbon;
+use Illuminate\Contracts\Auth\Authenticatable;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\File;
 
 class EmployeeService implements Import
 {
-    protected $headers = [
+    const REQUIRED_HEADERS = [
         'SCANNER ID',
         'FAMILY NAME',
         'GIVEN NAME',
-        'MIDDLE INITIAL',
-        'NAME EXTENSION',
         'REGULAR',
-        'OFFICE',
-        'ACTIVE',
     ];
+
+    private string $error = '';
+
+    protected $header = [];
 
     public function __construct(
         private Repository $repository
     ) { }
 
+    public function validate(UploadedFile $file): bool
+    {
+        $header = collect(self::REQUIRED_HEADERS)->every(fn ($header) => in_array($header, $this->headers((string) File::lines($file)->first())));
+
+        if ($header) {
+            $this->error = 'FILE UNSUPPORTED.';
+
+            return false;
+        }
+
+        $id = File::lines($file)->skip(1)->filter()->map(fn ($e) => str_getcsv($e)[$this->headers((string) File::lines($file)->first())['SCANNER ID']])->duplicates();
+
+        if ($id->isNotEmpty()) {
+            $this->error = "DUPLICATE IDS DETECTED: {$id->values()->toJSON()}. PLEASE CHECK AGAIN.";
+
+            return false;
+        }
+
+        return true;
+    }
+
+    public function error(): string
+    {
+        return $this->error;
+    }
+
     public function parse(UploadedFile $file): void
     {
-        $this->truncate();
+        $this->repository->truncate(fn ($query) => $query->whereUserId(auth()->id())->delete());
 
         File::lines($file)
             ->skip(1)
             ->filter()
-            ->map(fn($e) => str($e)->explode(','))
-            ->map(fn ($e) => $this->repository->transformImportData($e->toArray(), $this->headers((string) File::lines($file)->first())))
+            ->map(fn($e) => str_getcsv($e))
+            ->map(fn ($e) => $this->repository->transformImportData($e, $this->headers((string) File::lines($file)->first())))
             ->chunk(1000)
             ->each(fn ($e) => $this->repository->insert($e->toArray()));
+
+        event(new EmployeesImported(auth()->user(), $file));
     }
 
     public function headers(string $line): array
@@ -44,17 +76,14 @@ class EmployeeService implements Import
 
     public function all()
     {
-        return $this->repository
-            ->get(true)
-            ->where('user_id', auth()->id())
-            ->get();
+        return $this->repository->query()->whereUserId(auth()->id())->get();
     }
 
     public function offices()
     {
         return $this->repository
-            ->get(true)
-            ->where('user_id', auth()->id())
+            ->query()
+            ->whereUserId(auth()->id())
             ->select(['office', 'name'])
             ->get()
             ->map
@@ -65,17 +94,49 @@ class EmployeeService implements Import
             ->values();
     }
 
-    public function truncate()
-    {
-        return $this->repository->get(true)->where('user_id', auth()->id())->delete();
-    }
-
     public function update(string $id, array $payload)
     {
         $this->repository->update($id = explode(',', $id), [
             'user_id' => auth()->id(),
             ...$payload,
-            'nameToJSON' => true,
-        ], count($id) > 1 ? collect(request()->only('office', 'regular', 'active'))->filter(fn ($e) => $e == '*')->keys()->push('biometrics_id', 'name', 'user_id')->toArray() : []);
+            'nameToJSON' => 1,
+        ], count($id) > 1
+            ? collect(request()->only('office', 'regular', 'active'))
+                ->filter(fn ($e) => $e == '*')
+                ->keys()->push('biometrics_id', 'name', 'user_id')
+                ->toArray()
+            : []
+        );
+    }
+
+    public function markInactive(Authenticatable $user)
+    {
+        $user->employees()->active()->whereDoesntHave('mainLogs', function ($q) {
+            $q->whereDate('time', '>', today()->subMonth()->startOfMonth());
+        })->update(['active' => 0]);
+    }
+
+    public function markActive(Authenticatable $user)
+    {
+        $user->employees()->active(0)->whereHas('mainLogs', function ($q) {
+            $q->whereDate('time', '<', today()->subMonth()->startOfMonth());
+        })->update(['active' => 1]);
+    }
+
+    public function loadLogs(Collection|array $employees, ?Carbon $from = null, ?Carbon $to = null): Collection
+    {
+        $from ??= today()->startOfMonth();
+
+        $to ??= today()->endOfMonth();
+
+        if (is_array($employees)) {
+            $employees = $this->repository->find($employees);
+        }
+
+        $employees->filter->backedUp->map->load(['backupLogs' => fn ($q) => $q->whereBetween('time', [$from, $to])]);
+
+        $employees->reject->backedUp->map->load(['mainLogs' => fn ($q) => $q->whereBetween('time', [$from, $to])]);
+
+        return $employees;
     }
 }
